refactor: deleted .idea and updated directory name
