--- .idea/copyright/Dustin_Maurer.xml
@@ -1,6 +1,0 @@
-<component name="CopyrightManager">
-  <copyright>
-    <option name="notice" value="Copyright (c) &amp;#36;originalComment.match(&quot;Copyright \(c\) (\d+)&quot;, 1, &quot;-&quot;, &quot;&amp;#36;today&quot;)&amp;#36;today&#10;Since: &amp;#36;today.format(&quot;MMMM&quot;) &amp;#36;today.year&#10;  &lt;p&gt;&#10;Author: &amp;#36;username&#10; &lt;p&gt;&#10;Name: &amp;#36;file.fileName&#10; &lt;p&gt;&#10;Project-name: &amp;#36;project.name&#10;" />
-    <option name="myName" value="Dustin Maurer" />
-  </copyright>
-</component>
