--- .idea/copyright/profiles_settings.xml
@@ -1,7 +1,0 @@
-<component name="CopyrightManager">
-  <settings>
-    <module2copyright>
-      <element module="Project Files" copyright="Dustin Maurer" />
-    </module2copyright>
-  </settings>
-</component>
