+++ src/main/java/assignment/one_simple_list/SimpleFilter.java
@@ -1,4 +1,4 @@
-package org.example;
+package assignment.one_simple_list;
 @FunctionalInterface
 public interface SimpleFilter {
 	/**
